Implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\AskQuestionRequest;
 use App\Models\Question;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionsController extends Controller
 {
@@ -17,8 +18,9 @@ class QuestionsController extends Controller
     {
         $questions = Question::with('user')->latest()->paginate(5);
         $str = "\Illuminate\Support\Str";
+        $auth = "\Illuminate\Support\Facades\Auth";
 
-        return view('questions.index', compact('questions', 'str'));
+        return view('questions.index', compact('questions', 'str', 'auth'));
     }
 
     /**
@@ -67,6 +69,9 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if (Gate::denies('update-question', $question)) {
+            abort(403, "Access denied");
+        }
         return view("questions.edit", compact('question'));
     }
 
@@ -79,8 +84,11 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
-        $question->update($request->only('title', 'body'));
+        if (Gate::denies('update-question', $question)) {
+            abort(403, "Access denied");
+        }
 
+        $question->update($request->only('title', 'body'));
         return redirect('/questions')->with('success', "Your question has been updated.");
     }
 
@@ -92,8 +100,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (Gate::denies('delete-question', $question)) {
+            abort(403, "Access denied");
+        }
         $question->delete();
-
         return redirect('/questions')->with('success', "Your question has been deleted.");
     }
 }
